+++ include/ht_renderer.h
@@ -53,7 +53,7 @@ namespace Hatchit {
             OPENGL,
             DIRECTX11,
             DIRECTX12,
-			VULKAN
+            VULKAN
         };
 
         struct RendererParams
@@ -61,37 +61,42 @@ namespace Hatchit {
             RendererType renderer;
             void*        window;
             Color        clearColor;
-			std::string  applicationName;
+            std::string  applicationName;
         };
 
 
         class HT_API IRenderer
         {
         public:
+            static IRenderer* GetRenderer();
+
             virtual ~IRenderer() { };
-			
-			/** Initialize the renderer
-			* \param params The paramaters to intialize this renderer with
-			*/
+            
+            /** Initialize the renderer
+            * \param params The paramaters to intialize this renderer with
+            */
             virtual bool VInitialize(const RendererParams& params) = 0;
-			///Shutdown the renderer
+            ///Shutdown the renderer
             virtual void VDeInitialize() = 0;
-			/** Resizes the the screen
-			* \param width The new width of the screen
-			* \param height The new height of the screen
-			*/
+            /** Resizes the the screen
+            * \param width The new width of the screen
+            * \param height The new height of the screen
+            */
             virtual void VResizeBuffers(uint32_t width, uint32_t height) = 0;
 
-			/** Sets the color that the screen will clear with
-			* \param color The Color to clear the screen with
-			*/
+            /** Sets the color that the screen will clear with
+            * \param color The Color to clear the screen with
+            */
             virtual void VSetClearColor(const Color& color) = 0;
-			/** Clears the screen with the given clear color
-			* \param args Arguments to describe which buffer you want to clear
-			*/
+            /** Clears the screen with the given clear color
+            * \param args Arguments to describe which buffer you want to clear
+            */
             virtual void VClearBuffer(ClearArgs args) = 0;
-			///Present a frame to the screen via a backbuffer
+            ///Present a frame to the screen via a backbuffer
             virtual void VPresent() = 0;
+
+        protected:
+            static IRenderer* Renderer;
         };
 
     }
