--- source/vulkan/ht_vkpipeline.cpp
@@ -1,266 +1,0 @@
-/**
-**    Hatchit Engine
-**    Copyright(c) 2015 Third-Degree
-**
-**    GNU Lesser General Public License
-**    This file may be used under the terms of the GNU Lesser
-**    General Public License version 3 as published by the Free
-**    Software Foundation and appearing in the file LICENSE.LGPLv3 included
-**    in the packaging of this file. Please review the following information
-**    to ensure the GNU Lesser General Public License requirements
-**    will be met: https://www.gnu.org/licenses/lgpl.html
-**
-**/
-
-#include <ht_vkpipeline.h>
-
-namespace Hatchit {
-
-    namespace Graphics {
-
-        VKPipeline::VKPipeline() 
-        {
-            //Init the shader slots to nullptr
-            for (int i = 0; i < 6; i++)
-                m_shaderSlots[i] = nullptr;
-        }
-        VKPipeline::~VKPipeline() {}
-
-        //If we wanted to allow users to control blending states
-        //void VSetColorBlendAttachments(ColorBlendState* colorBlendStates) override;
-
-        /* Set the rasterization state for this pipeline
-        * \param rasterState A struct containing rasterization options
-        */
-        void VKPipeline::VSetRasterState(const RasterizerState& rasterState)
-        {
-            VkPolygonMode polyMode;
-            VkCullModeFlagBits cullMode;
-
-            switch (rasterState.polygonMode)
-            {
-            case PolygonMode::FILL:
-                polyMode = VK_POLYGON_MODE_FILL;
-                break;
-            case PolygonMode::LINE:
-                polyMode = VK_POLYGON_MODE_LINE;
-                break;
-            case PolygonMode::POINT:
-                polyMode = VK_POLYGON_MODE_LINE;
-                break;
-            default:
-                polyMode = VK_POLYGON_MODE_FILL;
-                break;
-            }
-
-            switch (rasterState.cullMode)
-            {
-            case NONE:
-                cullMode = VK_CULL_MODE_NONE;
-                break;
-            case FRONT:
-                cullMode = VK_CULL_MODE_FRONT_BIT;
-                break;
-            case BACK:
-                cullMode = VK_CULL_MODE_BACK_BIT;
-                break;
-            case FRONT_AND_BACK:
-                cullMode = VK_CULL_MODE_FRONT_AND_BACK;
-                break;
-            }
-            
-            m_rasterizationState = {}; //default setup
-            m_rasterizationState.sType = VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO;
-            m_rasterizationState.pNext = nullptr;
-            m_rasterizationState.polygonMode = polyMode;
-            m_rasterizationState.cullMode = cullMode;
-            m_rasterizationState.frontFace = VK_FRONT_FACE_COUNTER_CLOCKWISE;
-            m_rasterizationState.depthClampEnable = rasterState.depthClampEnable;
-            m_rasterizationState.rasterizerDiscardEnable = rasterState.discardEnable;
-            m_rasterizationState.depthBiasEnable = VK_FALSE;
-        }
-
-        /* Set the multisampling state for this pipeline
-        * \param multiState A struct containing multisampling options
-        */
-        void VKPipeline::VSetMultisampleState(const MultisampleState& multiState)
-        {
-            VkSampleCountFlagBits sampleCount;
-
-            switch (sampleCount)
-            {
-            case SAMPLE_1_BIT:
-                sampleCount = VK_SAMPLE_COUNT_1_BIT;
-                break;
-            case SAMPLE_2_BIT:
-                sampleCount = VK_SAMPLE_COUNT_2_BIT;
-                break;
-            case SAMPLE_4_BIT:
-                sampleCount = VK_SAMPLE_COUNT_4_BIT;
-                break;
-            case SAMPLE_8_BIT:
-                sampleCount = VK_SAMPLE_COUNT_8_BIT;
-                break;
-            case SAMPLE_16_BIT:
-                sampleCount = VK_SAMPLE_COUNT_16_BIT;
-                break;
-            case SAMPLE_32_BIT:
-                sampleCount = VK_SAMPLE_COUNT_32_BIT;
-                break;
-            case SAMPLE_64_BIT:
-                sampleCount = VK_SAMPLE_COUNT_64_BIT;
-                break;
-            }
-
-            m_multisampleState = {};
-            m_multisampleState.sType = VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO;
-            m_multisampleState.pNext = nullptr;
-            m_multisampleState.pSampleMask = nullptr;
-            m_multisampleState.rasterizationSamples = sampleCount;
-            m_multisampleState.sampleShadingEnable = multiState.perSampleShading;
-            m_multisampleState.minSampleShading = multiState.minSamples;
-        }
-
-        /* Load a shader into a shader slot for the pipeline
-        * \param shaderSlot The slot that you want the shader in; vertex, fragment etc.
-        * \param shader A pointer to the shader that you want to load to the given shader slot
-        */
-        void VKPipeline::VLoadShader(ShaderSlot shaderSlot, IShader* shader)
-        {
-            VKShader* vkShader = (VKShader*)shader;
-
-            VkShaderStageFlagBits shaderType;
-
-            switch (shaderSlot)
-            {
-            case ShaderSlot::VERTEX:
-                shaderType = VK_SHADER_STAGE_VERTEX_BIT;
-                break;
-            case ShaderSlot::FRAGMENT:
-                shaderType = VK_SHADER_STAGE_FRAGMENT_BIT;
-                break;
-            case ShaderSlot::GEOMETRY:
-                shaderType = VK_SHADER_STAGE_GEOMETRY_BIT;
-                break;
-            case ShaderSlot::TESS_CONTROL:
-                shaderType = VK_SHADER_STAGE_TESSELLATION_CONTROL_BIT;
-                break;
-            case ShaderSlot::TESS_EVAL:
-                shaderType = VK_SHADER_STAGE_TESSELLATION_EVALUATION_BIT;
-                break;
-            case ShaderSlot::COMPUTE:
-                shaderType = VK_SHADER_STAGE_COMPUTE_BIT;
-                break;
-            }
-
-            VkPipelineShaderStageCreateInfo shaderStage = {};
-            shaderStage.sType = VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
-            shaderStage.stage = shaderType;
-            shaderStage.module = vkShader->GetShaderModule();
-            shaderStage.pName = "main";
-
-            m_shaderStages.push_back(shaderStage);
-        }
-
-        ///Have Vulkan create a pipeline with these settings
-        bool VKPipeline::VPrepare()
-        {
-            VkResult err;
-
-            //Vertex info state
-            VkPipelineVertexInputStateCreateInfo vertexInputState = {};
-            vertexInputState.sType = VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO;
-
-            //Topology
-            VkPipelineInputAssemblyStateCreateInfo inputAssemblyState = {};
-            inputAssemblyState.sType = VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO;
-            inputAssemblyState.topology = VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST;
-
-            //Only one blend attachment state; not using blend right now
-            VkPipelineColorBlendAttachmentState blendAttachmentState[1] = {};
-            blendAttachmentState[0].colorWriteMask = 0xf;
-            blendAttachmentState[0].blendEnable = VK_FALSE;
-
-            //Color blends and masks
-            VkPipelineColorBlendStateCreateInfo colorBlendState = {};
-            colorBlendState.sType = VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
-            colorBlendState.pAttachments = blendAttachmentState;
-            colorBlendState.attachmentCount = 1;
-
-            //Viewport
-            VkPipelineViewportStateCreateInfo viewportState = {};
-            viewportState.sType = VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO;
-            viewportState.viewportCount = 1;
-            viewportState.scissorCount = 1;
-
-            //Enable dynamic states
-            VkDynamicState dynamicStateEnables[] = { VK_DYNAMIC_STATE_VIEWPORT , VK_DYNAMIC_STATE_SCISSOR };
-
-            VkPipelineDynamicStateCreateInfo dynamicState = {};
-            dynamicState.sType = VK_STRUCTURE_TYPE_PIPELINE_DYNAMIC_STATE_CREATE_INFO;
-            dynamicState.pNext = nullptr;
-            dynamicState.pDynamicStates = dynamicStateEnables;
-            dynamicState.dynamicStateCount = 2;
-
-            //Depth and stencil states
-            VkPipelineDepthStencilStateCreateInfo depthStencilState = {};
-            depthStencilState.sType = VK_STRUCTURE_TYPE_PIPELINE_DEPTH_STENCIL_STATE_CREATE_INFO;
-            depthStencilState.pNext = &vertexInputState;
-            depthStencilState.depthTestEnable = VK_TRUE;
-            depthStencilState.depthWriteEnable = VK_TRUE;
-            depthStencilState.depthCompareOp = VK_COMPARE_OP_LESS_OR_EQUAL;
-            depthStencilState.depthBoundsTestEnable = VK_FALSE;
-            depthStencilState.back.failOp = VK_STENCIL_OP_KEEP;
-            depthStencilState.back.passOp = VK_STENCIL_OP_KEEP;
-            depthStencilState.back.compareOp = VK_COMPARE_OP_ALWAYS;
-            depthStencilState.stencilTestEnable = VK_FALSE;
-            depthStencilState.front = depthStencilState.back;
-
-            //Finalize pipeline
-            VkGraphicsPipelineCreateInfo pipelineInfo = {};
-            pipelineInfo.sType = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
-            //pipelineInfo.layout = m_pipelineLayout;
-            //pipelineInfo.renderPass = m_renderPass;
-            pipelineInfo.stageCount = 2;
-            pipelineInfo.pVertexInputState = &vertexInputState;
-            pipelineInfo.pInputAssemblyState = &inputAssemblyState;
-            pipelineInfo.pRasterizationState = &m_rasterizationState;
-            pipelineInfo.pColorBlendState = &colorBlendState;
-            pipelineInfo.pMultisampleState = &m_multisampleState;
-            pipelineInfo.pViewportState = &viewportState;
-            pipelineInfo.pDepthStencilState = &depthStencilState;
-            pipelineInfo.pStages = m_shaderStages.data();
-            pipelineInfo.pDynamicState = &dynamicState;
-
-            VkPipelineCacheCreateInfo pipelineCacheInfo = {};
-            pipelineCacheInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_CACHE_CREATE_INFO;
-            //err = vkCreatePipelineCache(m_device, &pipelineCacheInfo, nullptr, &m_pipelineCache);
-            assert(!err);
-            if (err != VK_SUCCESS)
-            {
-#ifdef _DEBUG
-                Core::DebugPrintF("VKRenderer::preparePipeline(): Failed to create pipeline cache\n");
-#endif
-                return false;
-            }
-
-            //err = vkCreateGraphicsPipelines(m_device, m_pipelineCache, 1, &pipelineInfo, nullptr, &m_pipeline);
-            assert(!err);
-            if (err != VK_SUCCESS)
-            {
-#ifdef _DEBUG
-                Core::DebugPrintF("VKRenderer::preparePipeline(): Failed to create pipeline\n");
-#endif
-                return false;
-            }
-
-            //Delete shader modules; they're in the pipeline and we don't need them anymore
-            //vkDestroyShaderModule(m_device, m_shaderStages[0].module, nullptr);
-            //vkDestroyShaderModule(m_device, m_shaderStages[1].module, nullptr);
-
-
-            return true;
-        }
-        
-    }
-}
